change: Rename VulnScoreDisplay to HostsVulnScoreDisplay
Name the module after its component.
