+++ src/web/pages/hosts/dashboard/HostsVulnScoreDisplay.jsx
