+++ src/web/pages/start/NewDashboardDialog.jsx
@@ -13,8 +13,8 @@ import {CertBundCvssDisplay} from 'web/pages/certbund/dashboard/CvssDisplay';
 import {CvesCreatedDisplay} from 'web/pages/cves/dashboard/CreatedDisplay';
 import {CvesSeverityClassDisplay} from 'web/pages/cves/dashboard/SeverityClassDisplay';
 import {HostsTopologyDisplay} from 'web/pages/hosts/dashboard/HostsTopologyDisplay';
+import {HostsVulnScoreDisplay} from 'web/pages/hosts/dashboard/HostsVulnScoreDisplay';
 import {HostsModifiedDisplay} from 'web/pages/hosts/dashboard/ModifiedDisplay';
-import {HostsVulnScoreDisplay} from 'web/pages/hosts/dashboard/VulnScoreDisplay';
 import {NvtsSeverityClassDisplay} from 'web/pages/nvts/dashboard/SeverityClassDisplay';
 import {OsSeverityClassDisplay} from 'web/pages/operatingsystems/dashboard/SeverityClassDisplay';
 import {OsVulnScoreDisplay} from 'web/pages/operatingsystems/dashboard/VulnScoreDisplay';
